+++ src/MoneroLocalData.ts
@@ -1,63 +1,75 @@
-import type { EdgeTransaction } from 'edge-core-js'
-
-import { currencyInfo } from './moneroInfo'
+import {
+  asArray,
+  asBoolean,
+  asCodec,
+  asEither,
+  asJSON,
+  asNull,
+  asNumber,
+  asObject,
+  asOptional,
+  asString,
+  asTuple,
+  Cleaner,
+  uncleaner
+} from 'cleaners'
+import type { EdgeTokenId } from 'edge-core-js'
 
 export const DATA_STORE_FILE = 'txEngineFolder/walletLocalData.json'
-export const PRIMARY_CURRENCY = currencyInfo.currencyCode
+export const PRIMARY_CURRENCY_TOKEN_ID = null
 
-export class MoneroLocalData {
-  blockHeight: number
-  lastAddressQueryHeight: number
-  lockedXmrBalance: string
-  nextNonce: string
-  hasLoggedIn: boolean
-  totalBalances: { [currencyCode: string]: string }
-  enabledTokens: string[]
-  transactionsObj: { [currencyCode: string]: EdgeTransaction[] }
-
-  constructor(jsonString: string | null) {
-    this.blockHeight = 0
-
-    const totalBalances: { [currencyCode: string]: string } = { XMR: '0' }
-    this.totalBalances = totalBalances
-
-    this.nextNonce = '0'
-
-    this.lastAddressQueryHeight = 0
-    this.lockedXmrBalance = '0'
-
-    // Dumb extra local var needed to make Flow happy
-    const transactionsObj: {
-      [currencyCode: string]: EdgeTransaction[]
-    } = {}
-    this.transactionsObj = transactionsObj
-
-    this.hasLoggedIn = false
-    this.enabledTokens = [PRIMARY_CURRENCY]
-    if (jsonString !== null) {
-      const data = JSON.parse(jsonString)
-
-      if (typeof data.blockHeight === 'number') {
-        this.blockHeight = data.blockHeight
-      }
-      if (typeof data.hasLoggedIn === 'boolean') {
-        this.hasLoggedIn = data.hasLoggedIn
-      }
-      if (typeof data.lastAddressQueryHeight === 'string') {
-        this.lastAddressQueryHeight = data.lastAddressQueryHeight
-      }
-      if (typeof data.lockedXmrBalance === 'string') {
-        this.lockedXmrBalance = data.lockedXmrBalance
-      }
-      if (typeof data.totalBalances !== 'undefined') {
-        this.totalBalances = data.totalBalances
-      }
-      if (typeof data.enabledTokens !== 'undefined') {
-        this.enabledTokens = data.enabledTokens
-      }
-      if (typeof data.transactionsObj !== 'undefined') {
-        this.transactionsObj = data.transactionsObj
-      }
-    }
-  }
+const asNotNull: Cleaner<unknown> = (value: unknown) => {
+  if (value == null) throw new Error('Expected EdgeTransaction')
+  return value
 }
+
+const asMap = <K, V>(
+  asKey: Cleaner<K>,
+  asValue: Cleaner<V>
+): Cleaner<Map<K, V>> => {
+  return asCodec(
+    value => {
+      const unknownEntries =
+        value instanceof Map ? Array.from(value.entries()) : value
+      // TODO: Merge "Expected a Map" to TypeError message
+      const entries = asArray(asTuple(asKey, asValue))(unknownEntries)
+      return new Map(entries)
+    },
+    map => {
+      const entries = map.entries()
+      const entriesArray = Array.from(entries)
+      return entriesArray
+    }
+  )
+}
+
+const asEdgeToken: Cleaner<EdgeTokenId> = asEither(asString, asNull)
+const asCompatibleEdgeToken: Cleaner<EdgeTokenId> = (value: unknown) => {
+  // Backwards compatibility when the token used to be currencyCode
+  if (value === 'XMR') return null
+  return asEdgeToken(value)
+}
+
+export type MoneroLocalData = ReturnType<typeof asMoneroLocalData>
+export const asMoneroLocalData = asJSON(
+  asObject({
+    blockHeight: asOptional(asNumber, 0),
+    hasLoggedIn: asOptional(asBoolean, false),
+    lastAddressQueryHeight: asOptional(asNumber, 0),
+    lockedXmrBalance: asOptional(asString, '0'),
+    nextNonce: asOptional(asString, '0'),
+    totalBalances: asOptional(
+      asMap(asEdgeToken, asString),
+      () => new Map([[null, '0']])
+    ),
+    enabledTokens: asOptional(asArray(asCompatibleEdgeToken), [
+      PRIMARY_CURRENCY_TOKEN_ID
+    ]),
+    transactionsObj: asOptional(
+      asMap(asEdgeToken, asArray(asNotNull)),
+      () => new Map<EdgeTokenId, unknown[]>()
+    )
+  })
+)
+
+export const wasMoneroLocalData = uncleaner(asMoneroLocalData)