+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- fixed: Upgrade edge-core-js to 2.26.0 to implemented new Seen Tx Checkpoint API
+
 # 1.3.2 (2024-11-19)
 
 - fixed: URI parsing and creation according to the correct format